+++ src/shared/qclass_lib_map.h
@@ -692,6 +692,6 @@ static const ClassInfoEntry qclass_lib_map[] = {
     { "QFormBuilder", "QtUiTools/formbuilder.h"},
     { "QUiLoader", "QtUiTools/quiloader.h"},
 };
-static const int qclass_lib_count = 686;
+static const qint16 qclass_lib_count = 686;
 
 #endif
